fix(portal): use correct per-currency decimal precision
Extract shared currency formatting to utils/currency.ts matching
backend precision (money.rs): USD=2, TWD=0, BTC=8, ETH=18→8, USDT=6.
Fixes TWD showing 2 decimal places instead of 0. Removes duplicate
formatBalance/formatAmount from Accounts and Transactions pages.
